+++ src/urisys/markpact/analyzer/__init__.py
@@ -1,4 +1,5 @@
 from .lint import run_lint
+from .format import ANALYZE_JSON_FORMAT, analyze_json_report, collect_analyze_issues
 from .report import analyze_markpact
 
-__all__ = ["analyze_markpact", "run_lint"]
+__all__ = ["analyze_markpact", "analyze_json_report", "collect_analyze_issues", "ANALYZE_JSON_FORMAT", "run_lint"]